Synchronize codes for SM7325R OPPO Reno 7 5G

+++ drivers/staging/rtl8192u/r8192U_core.c
@@ -3248,7 +3248,7 @@ static void rtl819x_update_rxcounts(struct r8192_priv *priv, u32 *TotalRxBcnNum,
 			     u32 *TotalRxDataNum)
 {
 	u16			SlotIndex;
-	u16			i;
+	u8			i;
 
 	*TotalRxBcnNum = 0;
 	*TotalRxDataNum = 0;
